requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
matplotlib==3.4.3
seaborn==0.11.2
jupyter==1.0.0
notebook==6.4.4
nbconvert==6.1.0
groq==0.4.0
python-dotenv==1.0.0 